+++ packages/core/package.json
@@ -18,7 +18,6 @@
     "@types/inflection": "^2.0.0",
     "@types/ini": "^4.1.1",
     "@types/jsdom": "^21.1.7",
-    "@types/mdast": "^4.0.4",
     "@types/mime-types": "^3.0.1",
     "@types/mustache": "^4.2.6",
     "@types/node": "catalog:",
@@ -96,7 +95,6 @@
     "magic-string": "^0.30.17",
     "mammoth": "^1.9.1",
     "mathjs": "^14.5.2",
-    "mdast": "^3.0.0",
     "merge-descriptors": "^2.0.0",
     "mime": "^4.0.7",
     "minimatch": "^10.0.3",
@@ -112,12 +110,6 @@
     "parse-diff": "^0.11.1",
     "pretty-bytes": "^7.0.0",
     "pretty-ms": "^9.2.0",
-    "remark": "^15.0.1",
-    "remark-frontmatter": "^5.0.0",
-    "remark-gfm": "^4.0.1",
-    "remark-github": "^12.0.0",
-    "remark-parse": "^11.0.0",
-    "remark-stringify": "^11.0.0",
     "sanitize-filename": "^1.6.3",
     "semver": "^7.7.2",
     "serialize-error": "^12.0.0",
@@ -131,7 +123,6 @@
     "turndown": "^7.2.0",
     "turndown-plugin-gfm": "^1.0.2",
     "undici": "^7.10.0",
-    "unified": "^11.0.5",
     "uuid": "^11.1.0",
     "yaml": "^2.8.0",
     "zod-to-json-schema": "^3.24.6"
