+++ packages/runtime/package.json
@@ -20,9 +20,20 @@
     "es-toolkit": "catalog:",
     "execa": "catalog:",
     "glob": "^11.0.3",
+    "@types/mdast": "^4.0.4",
+    "mdast": "^3.0.0",
     "memorystream": "^0.3.1",
     "playwright": "^1.53.1",
+    "remark": "^15.0.1",
+    "remark-directive": "^4.0.0",
+    "remark-frontmatter": "^5.0.0",
+    "remark-gfm": "^4.0.1",
+    "remark-github": "^12.0.0",
+    "remark-math": "^6.0.0",
+    "remark-parse": "^11.0.0",
+    "remark-stringify": "^11.0.0",
     "tslib": "catalog:",
+    "unist-util-visit": "^5.0.0",
     "zod": "^3.25.67"
   },
   "devDependencies": {
