+++ packages/runtime/src/index.ts
@@ -14,3 +14,4 @@ export * from "./filetree.js";
 export * from "./markdownifypdf.js";
 export * from "./runtime.js";
 export * from "./extras.js";
+export * from "./mdast.js";
